remove compress modules
