+++ sha2/src/sha256/soft.rs
