+++ sha2/src/sha256/x86.rs
